Add files via upload
A toy factory is packing toys into boxes. Each box can contain either a toy (represented by an integer value) or be empty (represented by 0). Write a function to move all empty boxes to the end of the array while maintaining the order of the toys, and also return the count of empty boxes.
Sample Input :
Ex : Boxes = [1,0,2,0,3,4,0,5]